Move PageXmlizer and WikiPageProperties to fitnesse.wiki.fs.
