+++ test/fitnesse/wiki/fs/PageXmlizerTest.java
@@ -1,6 +1,6 @@
 // Copyright (C) 2003-2009 by Object Mentor, Inc. All rights reserved.
 // Released under the terms of the CPL Common Public License version 1.0.
-package fitnesse.wiki;
+package fitnesse.wiki.fs;
 
 import java.text.DateFormat;
 import java.util.Date;
@@ -11,7 +11,10 @@ import org.junit.Test;
 import org.w3c.dom.Document;
 
 import fitnesse.util.XmlUtil;
+import fitnesse.wiki.*;
 import fitnesse.wiki.fs.InMemoryPage;
+import fitnesse.wiki.fs.PageXmlizer;
+import fitnesse.wiki.fs.WikiPageProperties;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
